# services/access/internal/api/openapi_gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.2 DO NOT EDIT.
package api

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	ScannerDeviceTokenScopes scannerDeviceTokenContextKey = "ScannerDeviceToken.Scopes"
)

// Defines values for ReconcileResultResult.
const (
	ReconcileResultResultConflict ReconcileResultResult = "conflict"
	ReconcileResultResultRecorded ReconcileResultResult = "recorded"
	ReconcileResultResultRejected ReconcileResultResult = "rejected"
	ReconcileResultResultSynced   ReconcileResultResult = "synced"
)

// Valid indicates whether the value is a known member of the ReconcileResultResult enum.
func (e ReconcileResultResult) Valid() bool {
	switch e {
	case ReconcileResultResultConflict:
		return true
	case ReconcileResultResultRecorded:
		return true
	case ReconcileResultResultRejected:
		return true
	case ReconcileResultResultSynced:
		return true
	default:
		return false
	}
}

// Defines values for ScanAcceptedDecision.
const (
	Accepted ScanAcceptedDecision = "accepted"
)

// Valid indicates whether the value is a known member of the ScanAcceptedDecision enum.
func (e ScanAcceptedDecision) Valid() bool {
	switch e {
	case Accepted:
		return true
	default:
		return false
	}
}

// Defines values for ScanRejectedDecision.
const (
	ScanRejectedDecisionRejected ScanRejectedDecision = "rejected"
)

// Valid indicates whether the value is a known member of the ScanRejectedDecision enum.
func (e ScanRejectedDecision) Valid() bool {
	switch e {
	case ScanRejectedDecisionRejected:
		return true
	default:
		return false
	}
}

// Defines values for ScanRequestDirection.
const (
	Entry ScanRequestDirection = "entry"
	Exit  ScanRequestDirection = "exit"
)

// Valid indicates whether the value is a known member of the ScanRequestDirection enum.
func (e ScanRequestDirection) Valid() bool {
	switch e {
	case Entry:
		return true
	case Exit:
		return true
	default:
		return false
	}
}

// Error defines model for Error.
type Error struct {
	Error string `json:"error"`
}

// LifecycleEvent defines model for LifecycleEvent.
type LifecycleEvent struct {
	Id         openapi_types.UUID `json:"id"`
	OccurredAt time.Time          `json:"occurred_at"`
	Sequence   *int64             `json:"sequence,omitempty"`
	Type       string             `json:"type"`
}

// ReconcileOccurrence defines model for ReconcileOccurrence.
type ReconcileOccurrence struct {
	EventType    *string `json:"event_type,omitempty"`
	OccurredAt   string  `json:"occurred_at"`
	OccurrenceId string  `json:"occurrence_id"`
	QrPayload    string  `json:"qr_payload"`
}

// ReconcileRequest defines model for ReconcileRequest.
type ReconcileRequest struct {
	Occurrences []ReconcileOccurrence `json:"occurrences"`
}

// ReconcileResponse defines model for ReconcileResponse.
type ReconcileResponse struct {
	Results []ReconcileResult `json:"results"`
}

// ReconcileResult defines model for ReconcileResult.
type ReconcileResult struct {
	OccurredAt   *time.Time            `json:"occurred_at,omitempty"`
	OccurrenceId string                `json:"occurrence_id"`
	Result       ReconcileResultResult `json:"result"`
	SkewFlagged  *bool                 `json:"skew_flagged,omitempty"`
}

// ReconcileResultResult defines model for ReconcileResult.Result.
type ReconcileResultResult string

// ScanAccepted defines model for ScanAccepted.
type ScanAccepted struct {
	Decision  ScanAcceptedDecision `json:"decision"`
	Replay    *bool                `json:"replay,omitempty"`
	ScannedAt time.Time            `json:"scanned_at"`
}

// ScanAcceptedDecision defines model for ScanAccepted.Decision.
type ScanAcceptedDecision string

// ScanRejected defines model for ScanRejected.
type ScanRejected struct {
	Decision       ScanRejectedDecision `json:"decision"`
	OriginalScanAt *time.Time           `json:"original_scan_at,omitempty"`
	Reason         string               `json:"reason"`
}

// ScanRejectedDecision defines model for ScanRejected.Decision.
type ScanRejectedDecision string

// ScanRequest defines model for ScanRequest.
type ScanRequest struct {
	Direction    *ScanRequestDirection `json:"direction,omitempty"`
	OccurredAt   *string               `json:"occurred_at,omitempty"`
	OccurrenceId *string               `json:"occurrence_id,omitempty"`
	QrPayload    string                `json:"qr_payload"`
}

// ScanRequestDirection defines model for ScanRequest.Direction.
type ScanRequestDirection string

// Ticket defines model for Ticket.
type Ticket struct {
	History   []LifecycleEvent   `json:"history"`
	IssuedAt  time.Time          `json:"issued_at"`
	QrPayload string             `json:"qr_payload"`
	QrUrl     string             `json:"qr_url"`
	TicketId  openapi_types.UUID `json:"ticket_id"`
}

// TicketBundle defines model for TicketBundle.
type TicketBundle struct {
	OrderRef openapi_types.UUID `json:"order_ref"`
	Tickets  []Ticket           `json:"tickets"`
}

// TicketRefund defines model for TicketRefund.
type TicketRefund struct {
	Replay    bool                 `json:"replay"`
	TicketIds []openapi_types.UUID `json:"ticket_ids"`
}

// TicketRefundCreate defines model for TicketRefundCreate.
type TicketRefundCreate struct {
	OrganizerId openapi_types.UUID `json:"organizer_id"`
	Quantity    int                `json:"quantity"`
	RefundId    openapi_types.UUID `json:"refund_id"`
}

// Ref defines model for Ref.
type Ref = openapi_types.UUID

// scannerDeviceTokenContextKey is the context key for ScannerDeviceToken security scheme
type scannerDeviceTokenContextKey string

// GetTicketQRParams defines parameters for GetTicketQR.
type GetTicketQRParams struct {
	// Exp Signature expiry, as Unix seconds.
	Exp string `form:"exp" json:"exp"`

	// Sig base64url HMAC over the order ref, the ticket id and the expiry.
	Sig string `form:"sig" json:"sig"`
}

// RefundTicketsJSONRequestBody defines body for RefundTickets for application/json ContentType.
type RefundTicketsJSONRequestBody = TicketRefundCreate

// ScanTicketJSONRequestBody defines body for ScanTicket for application/json ContentType.
type ScanTicketJSONRequestBody = ScanRequest

// ReconcileScansJSONRequestBody defines body for ReconcileScans for application/json ContentType.
type ReconcileScansJSONRequestBody = ReconcileRequest