# services/commerce/api/openapi.yaml
openapi: 3.0.3
info: {title: Commerce API, version: 1.0.0}
paths:
...
  /internal/orders/{id}/refunds:
    post:
      operationId: refundOrder
      summary: Staff refund of all or part of a completed order (money leg only)
      # Post-purchase refund (TKT-156, ADR-037). The order's checkout status stays
      # `completed` — a refunded purchase is still a purchase that happened, and
      # `orders.status='refunded'` already means something else (a compensated FAILED
      # checkout, ADR-016). Refund state is reported on its own dimension.
      # Returning the inventory capacity and voiding the tickets is TKT-157.
      parameters: [{$ref: '#/components/parameters/Id'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RefundCreate'}}}}
      responses:
        '200': {description: Refunded (or replay of a completed refund), content: {application/json: {schema: {$ref: '#/components/schemas/Refund'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '500': {$ref: '#/components/responses/Error'}
        '502': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
components:
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 1, maxLength: 200}}
...
  responses:
    Error: {description: Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error: {type: string}
        code: {type: string, enum: [seat_taken, orphaned_seats], description: 'Machine-readable refusal reason. `seat_taken`: a seated reservation lost seats to a competing claimant (TKT-173). `orphaned_seats`: the selection would strand free seats with no free neighbour (ADR-041, TKT-182).'}
        seat_identities:
          type: array
          description: >
            With `seat_taken`: the requested seats another buyer already holds,
            sorted — a SUBSET of the request, forwarded verbatim from the inventory
            transaction that arbitrated. With `orphaned_seats`: the FREE seats the
            selection would strand — seats the buyer did NOT request, so the subset
            rule deliberately does not apply to them and a picker must keep them
            selectable, since adding one is the repair. Never synthesised by
            commerce: a refusal without a usable identity list is a 502, not a guess.
          items: {type: string, minLength: 1, maxLength: 200}
...
    RefundCreate:
      type: object
      additionalProperties: false
      required: [organizer_id, quantity, actor, reason]
      properties:
        organizer_id: {type: string, format: uuid}
        quantity: {type: integer, minimum: 1, maximum: 50}
        # Staff attribution, mirroring OperationalConversionCreate: an operation that
        # moves money cannot be less attributable than one that moves a seat.
        actor: {type: string, minLength: 1, maxLength: 200}
        reason: {type: string, minLength: 1, maxLength: 500}
    Refund:
      type: object
      additionalProperties: false
      required: [refund_id, order_id, quantity, amount, currency, refund_status, refunded_quantity, refunded_amount, replay, tickets_voided, capacity_returned]
      properties:
        refund_id: {type: string, format: uuid}
        order_id: {type: string, format: uuid}
        quantity: {type: integer}
        amount: {type: integer, format: int64}
        currency: {type: string}
        refund_status: {type: string, enum: [none, partial, full]}
        refunded_quantity: {type: integer}
        refunded_amount: {type: integer, format: int64}
        replay: {type: boolean}
        # False means the tickets are still valid at the gate — the money moved but
        # the reversal is outstanding. Required, not optional: a caller that cannot
        # tell "voided" from "we did not get to it" will assume the first (TKT-157).
        tickets_voided: {type: boolean}
        # False means the seat is still counted as sold. Required for the same reason as
        # tickets_voided: a caller that cannot tell "returned" from "we did not get to it"
        # will assume the first. A partial refund of a SEATED order stays false forever —
        # no subset of seats can be identified (TKT-164).
        capacity_returned: {type: boolean}
...